authenticated admin endpoint and made it a role based entry

+++ journalApp/src/main/java/com/salAce/journalApp/service/JournalEntryService.java
@@ -3,6 +3,7 @@ package com.salAce.journalApp.service;
 import com.salAce.journalApp.entity.JournalEntry;
 import com.salAce.journalApp.entity.User;
 import com.salAce.journalApp.repo.JournalEntryRepo;
+import lombok.extern.slf4j.Slf4j;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -12,6 +13,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
+@Slf4j
 @Component
 public class JournalEntryService {
 
@@ -29,11 +31,11 @@ public class JournalEntryService {
 
                 JournalEntry saved =   journalEntryRepo.save(journalEntry) ;
                 foundUser.getJournalEntries().add(saved) ;
-//                foundUser.setUserName(null);
-                userEntryService.updateJournalEntries(foundUser) ;
-//                userEntryService.saveEntry(foundUser) ;
+
+                userEntryService.saveEntry(foundUser) ;
+
             } catch(Exception e){
-//                System.out.println(e.getMessage());
+
                 throw new Exception("generated exception" , e) ;
             }
 
@@ -53,17 +55,35 @@ public class JournalEntryService {
             return journalEntryRepo.findById(id) ;
 
         }
+@Transactional
+        public boolean deleteById(ObjectId id, String userName) {
+        boolean removed = false ;
+    try {
+        User foundUser = userEntryService.findByUserName(userName) ;  // removeIf returns boolean
+        removed =   foundUser.getJournalEntries().removeIf(x -> x.getId().equals(id)) ;
 
-        public void deleteById(ObjectId id, String userName) {
-
-            User foundUser = userEntryService.findByUserName(userName) ;
-            foundUser.getJournalEntries().removeIf(x -> x.getId().equals(id)) ;
-
-            // user ko find kro fir uska journal entries nikaalo fir usme us id ko search kro if present then delete it
-
+        // user ko find kro fir uska journal entries nikaalo fir usme us id ko search kro if present then delete it
+        if(removed){
+            userEntryService.saveEntry(foundUser); // save the foundUser i.e remove the id of that journal....
             journalEntryRepo.deleteById(id) ;
+            System.out.println("the journal was removed and user was updated");
+            userEntryService.saveEntry(foundUser);
+        }
+    }catch(Exception e ) {
+
+        System.out.println(e);
+        throw new RuntimeException("an error has occured while deleting the entry " ,e );
+
+
+    }
+    return removed ;
+
 
         }
+//        public List<JournalEntry> findByUserName(String userName) {
+//
+//
+//        }
 
     }
